requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.3
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
joblib==1.1.0
mypy-extensions==0.4.3
numpy==1.21.2
pathspec==0.9.0
platformdirs==2.4.0
pytz==2021.3
regex==2021.10.8
Sastrawi==1.0.1
scikit-learn==0.24.2
scipy==1.7.1
sqlparse==0.4.2
threadpoolctl==3.0.0
tomli==1.2.1
typing-extensions==3.10.0.2
whitenoise==5.3.0
